adhesion en proceso

+++ ComprasCartonesLGP.Entities/AdhesionCbu.cs
@@ -15,7 +15,7 @@ namespace ComprasCartonesLGP.Entities
         public string external_reference { get; set; }//external_reference	String	Este atributo se puede utilizar como referencia para identificar la Adhesión y sincronizar con tus sistemas de backend el origen de la operación. Algunos valores comúnmente utilizados son: ID de Cliente, DNI, CUIT, ID de venta o Nro. de Factura entre otros.
         public string adhesion_holder_name { get; set; }//adhesion_holder_name	String	Nombre del titular del servicio que se debitará.
         public string email { get; set; }//email	String	Email del del titular de la cuenta bancaria.
-        public int cbu_holder_id_number { get; set; }//cbu_holder_id_number	Integer	CUIT/CUIL del títular de la cuenta bancaria.
+        public long cbu_holder_id_number { get; set; }//cbu_holder_id_number	Integer	CUIT/CUIL del títular de la cuenta bancaria.
         public string cbu_holder_name { get; set; }//cbu_holder_name	String	Nombre del titular de la cuenta bancaria.
         public string cbu_number { get; set; }//cbu_number	String	Número de CBU de la cuenta bancaria en la que se ejecutarán los débitos.
         public string bank { get; set; }//bank	String	Nombre de la entidad bancaria a la que corresponde el número de CBU.
